Added searchbar start
I removed dropdown from this branch. I have followed through the basics of the tutorial and created a basic layout. Problems: We need to hide it at appropriate times so it doesn't cover the screen ideally.(Or maybe we put it at the bottom). Also table cell selection looks junky and could use work figuring out how to pull data from a selection and update the bar.